refactoring
AI-Session-Id: ef825fac-c03e-4845-815f-4e2c04224f67
AI-Tool: claude-code
AI-Model: unknown
